# Introduce the `small-hash` feature for `bitcoin_hashes`