--- a/clippy.toml
+++ b/clippy.toml
@@ -1,1 +1,2 @@
 msrv = "1.48.0"
+too-many-arguments-threshold = 13