--- a/hashes/Cargo.toml
+++ b/hashes/Cargo.toml
@@ -19,6 +19,8 @@ alloc = ["internals/alloc", "hex/alloc"]
 serde-std = ["serde/std"]
 # If you want I/O you must enable either "std" or "core2".
 core2 = ["actual-core2", "hex/core2"]
+# Smaller (but slower) implementation of sha256, sha512 and ripemd160
+small-hash = []
 
 [package.metadata.docs.rs]
 all-features = true