@apoelstra 2023-08-11
Can you edit `clippy.toml` to set `too-many-arguments-threshold` to 9 (or whatever value you need).