@apoelstra 2023-08-12
I agree that it's a little weird to use a cargo feature, but also think it's the right choice, for the same reasons as the rust-secp `lowmemory` feature. That is, it has zero effect on functionality, and if some dep wants this on, it doesn't (really) make sense for any other dep to be able to turn it off.