@apoelstra 2023-08-14
@tcharding I suspect that constfns are just as fast, but the reason that they (can be) smaller is that macros are **always** inlined whereas constfns might not be. Inlining is faster because it avoids function call overhead.

Also +1 to custom formatting.