@tcharding 2023-08-14
So I tried to read the [dragon book](https://suif.stanford.edu/dragonbook/) a couple of times before and never made it past the first chapters. Anyone got an suggestions on books/sites to learn about compiler optimisations without making my eyes bleed?