@tcharding 2023-08-14
With green CI I'm happy to ack this. Its basically just swapping macros for const functions, it did take me a while to work that out though :)