@apoelstra 2023-08-16
> Do we want this in the upcoming `hashes` release?

Yeah, let's do it.