@apoelstra 2023-08-16
Nice! I suspect `round` could also be a `fn` but since it can't be a constfn I am less confident that the compiler will inline it. Maybe we could add `cfg_attr(feature = "small_hashes", inline(never))` on it (and a corresponding `inline(always)` in the other case)?

But in the interest of not making you do a ton of iterations on this, I'm happy to ACK this as is. But if you want to try this, that'd be great :).